# build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Disk": {
				"properties": {
					"index": {
						"type": "number",
						"format": "double"
					},
					"model": {
						"type": "string"
					},
					"serialNumber": {
						"type": "string"
					},
					"firmwareRevision": {
						"type": "string"
					},
					"sizeBytes": {
						"type": "number",
						"format": "double"
					},
					"temperature": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"lastErrorCode": {
						"type": "string",
						"nullable": true
					},
					"errorCleared": {
						"type": "string",
						"nullable": true
					},
					"errorDescription": {
						"type": "string",
						"nullable": true
					},
					"errorMethodology": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"index",
					"model",
					"serialNumber",
					"firmwareRevision",
					"sizeBytes",
					"temperature",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partition": {
				"properties": {
					"diskIndex": {
						"type": "number",
						"format": "double"
					},
					"index": {
						"type": "number",
						"format": "double"
					},
					"bootable": {
						"type": "boolean"
					},
					"primaryPartition": {
						"type": "boolean"
					},
					"volumeName": {
						"type": "string"
					},
					"sizeBytes": {
						"type": "number",
						"format": "double"
					},
					"freeSpaceBytes": {
						"type": "number",
						"format": "double"
					},
					"usagePercentage": {
						"type": "number",
						"format": "double"
					},
					"volumeLetter": {
						"type": "string",
						"nullable": true
					},
					"fileSystem": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"diskIndex",
					"index",
					"bootable",
					"primaryPartition",
					"volumeName",
					"sizeBytes",
					"freeSpaceBytes",
					"usagePercentage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IStorageInformation": {
				"properties": {
					"machineId": {
						"type": "string"
					},
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"disks": {
						"items": {
							"$ref": "#/components/schemas/Disk"
						},
						"type": "array"
					},
					"partitions": {
						"items": {
							"$ref": "#/components/schemas/Partition"
						},
						"type": "array"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"machineId",
					"timestamp",
					"disks",
					"partitions"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_StorageInformation.Exclude_keyofStorageInformation.machineId__": {
				"properties": {
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"disks": {
						"items": {
							"$ref": "#/components/schemas/Disk"
						},
						"type": "array"
					},
					"partitions": {
						"items": {
							"$ref": "#/components/schemas/Partition"
						},
						"type": "array"
					}
				},
				"required": [
					"timestamp",
					"disks",
					"partitions"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_StorageInformation.machineId_": {
				"$ref": "#/components/schemas/Pick_StorageInformation.Exclude_keyofStorageInformation.machineId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"INetworkInformation": {
				"properties": {
					"machineId": {
						"type": "string"
					},
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"bytesSent": {
						"type": "number",
						"format": "double"
					},
					"bytesReceived": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"speed": {
						"type": "string"
					},
					"ipAddress": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"machineId",
					"timestamp",
					"name",
					"bytesSent",
					"bytesReceived",
					"type",
					"status",
					"speed",
					"ipAddress"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_NetworkInformation.Exclude_keyofNetworkInformation.machineId__": {
				"properties": {
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"bytesSent": {
						"type": "number",
						"format": "double"
					},
					"bytesReceived": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"speed": {
						"type": "string"
					},
					"ipAddress": {
						"type": "string"
					}
				},
				"required": [
					"timestamp",
					"name",
					"bytesSent",
					"bytesReceived",
					"type",
					"status",
					"speed",
					"ipAddress"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_NetworkInformation.machineId_": {
				"$ref": "#/components/schemas/Pick_NetworkInformation.Exclude_keyofNetworkInformation.machineId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"CpuCore": {
				"properties": {
					"coreId": {
						"type": "number",
						"format": "double"
					},
					"temperature": {
						"type": "number",
						"format": "double"
					},
					"load": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"coreId",
					"temperature",
					"load"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IMemoryInformation": {
				"properties": {
					"machineId": {
						"type": "string"
					},
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"usedMemoryGB": {
						"type": "number",
						"format": "double"
					},
					"availableMemoryGB": {
						"type": "number",
						"format": "double"
					},
					"totalMemoryGB": {
						"type": "number",
						"format": "double"
					},
					"usagePercentage": {
						"type": "number",
						"format": "double"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"machineId",
					"timestamp",
					"usedMemoryGB",
					"availableMemoryGB",
					"totalMemoryGB",
					"usagePercentage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_MemoryInformation.Exclude_keyofMemoryInformation.machineId__": {
				"properties": {
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"usedMemoryGB": {
						"type": "number",
						"format": "double"
					},
					"availableMemoryGB": {
						"type": "number",
						"format": "double"
					},
					"totalMemoryGB": {
						"type": "number",
						"format": "double"
					},
					"usagePercentage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"timestamp",
					"usedMemoryGB",
					"availableMemoryGB",
					"totalMemoryGB",
					"usagePercentage"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_MemoryInformation.machineId_": {
				"$ref": "#/components/schemas/Pick_MemoryInformation.Exclude_keyofMemoryInformation.machineId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"ICpuInformation": {
				"properties": {
					"machineId": {
						"type": "string"
					},
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"temperature": {
						"type": "number",
						"format": "double"
					},
					"load": {
						"type": "number",
						"format": "double"
					},
					"power": {
						"type": "number",
						"format": "double"
					},
					"cores": {
						"items": {
							"$ref": "#/components/schemas/CpuCore"
						},
						"type": "array"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"machineId",
					"timestamp",
					"name",
					"temperature",
					"load",
					"power",
					"cores"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_CpuInformation.Exclude_keyofCpuInformation.machineId__": {
				"properties": {
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"temperature": {
						"type": "number",
						"format": "double"
					},
					"load": {
						"type": "number",
						"format": "double"
					},
					"power": {
						"type": "number",
						"format": "double"
					},
					"cores": {
						"items": {
							"$ref": "#/components/schemas/CpuCore"
						},
						"type": "array"
					}
				},
				"required": [
					"timestamp",
					"name",
					"temperature",
					"load",
					"power",
					"cores"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_CpuInformation.machineId_": {
				"$ref": "#/components/schemas/Pick_CpuInformation.Exclude_keyofCpuInformation.machineId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IAntivirusInformation": {
				"properties": {
					"machineId": {
						"type": "string"
					},
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"enabled": {
						"type": "boolean"
					},
					"lastQuickScan": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"lastFullScan": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"hasThreats": {
						"type": "boolean",
						"nullable": true
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"machineId",
					"timestamp",
					"name",
					"enabled"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_AntivirusInformation.Exclude_keyofAntivirusInformation.machineId__": {
				"properties": {
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"name": {
						"type": "string"
					},
					"enabled": {
						"type": "boolean"
					},
					"lastQuickScan": {
						"type": "string",
						"format": "date-time"
					},
					"lastFullScan": {
						"type": "string",
						"format": "date-time"
					},
					"hasThreats": {
						"type": "boolean"
					}
				},
				"required": [
					"timestamp",
					"name",
					"enabled"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_AntivirusInformation.machineId_": {
				"$ref": "#/components/schemas/Pick_AntivirusInformation.Exclude_keyofAntivirusInformation.machineId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
...
			"IMachine": {
				"properties": {
					"machineId": {
						"type": "string"
					},
					"computerName": {
						"type": "string"
					},
					"operatingSystem": {
						"type": "string"
					},
					"operatingSystemVersion": {
						"type": "string"
					},
					"operatingSystemArchitecture": {
						"type": "string"
					},
					"operatingSystemDescription": {
						"type": "string"
					},
					"domainName": {
						"type": "string"
					},
					"workgroup": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"lastSeen": {
						"type": "string",
						"format": "date-time"
					},
					"description": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"owner": {
						"type": "string"
					},
					"organizationId": {
						"type": "string"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"machineId",
					"computerName",
					"operatingSystem",
					"operatingSystemVersion",
					"operatingSystemArchitecture",
					"operatingSystemDescription",
					"isActive"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Machine": {
				"properties": {
					"machineId": {
						"type": "string"
					},
					"computerName": {
						"type": "string"
					},
					"operatingSystem": {
						"type": "string"
					},
					"operatingSystemVersion": {
						"type": "string"
					},
					"operatingSystemArchitecture": {
						"type": "string"
					},
					"operatingSystemDescription": {
						"type": "string"
					},
					"domainName": {
						"type": "string"
					},
					"workgroup": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"lastSeen": {
						"type": "string",
						"format": "date-time"
					},
					"description": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"owner": {
						"type": "string"
					},
					"organizationId": {
						"type": "string"
					}
				},
				"required": [
					"machineId",
					"computerName",
					"operatingSystem",
					"operatingSystemVersion",
					"operatingSystemArchitecture",
					"operatingSystemDescription",
					"isActive"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "X-Token",
				"in": "header",
				"description": "JWT token sent via X-Token header. Example: \"X-Token: {token}\""
			}
		}
	},
	"info": {
		"title": "api-tor",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/monitoring/storage-information/machine/{machineId}": {
			"get": {
				"operationId": "GetStorageInformationByMachineId",
				"responses": {
					"200": {
						"description": "List",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IStorageInformation"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Storage Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateStorageInformation",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "string"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Storage Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_StorageInformation.machineId_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteStorageInformationByMachineId",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"deletedCount": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"deletedCount"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Storage Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/monitoring/network-information/machine/{machineId}": {
			"get": {
				"operationId": "GetNetworkInformationByMachineId",
				"responses": {
					"200": {
						"description": "List",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/INetworkInformation"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Network Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateNetworkInformation",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "string"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Network Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_NetworkInformation.machineId_"
							}
						}
					}
				}
			}
		},
...
		"/monitoring/memory-information/machine/{machineId}": {
			"get": {
				"operationId": "GetMemoryInformationByMachineId",
				"responses": {
					"200": {
						"description": "List",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IMemoryInformation"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Memory Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateMemoryInformation",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "string"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Memory Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_MemoryInformation.machineId_"
							}
						}
					}
				}
			}
		},
...
		"/monitoring/cpu-information/machine/{machineId}": {
			"get": {
				"operationId": "GetCpuInformationByMachineId",
				"responses": {
					"200": {
						"description": "List",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ICpuInformation"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Get all CPU information for a machine",
				"tags": [
					"CPU Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateCpuInformation",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "string"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Create single CPU information record",
				"tags": [
					"CPU Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_CpuInformation.machineId_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteCpuInformationByMachineId",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"deletedCount": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"deletedCount"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Delete all CPU information for a machine",
				"tags": [
					"CPU Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/monitoring/antivirus-information/machine/{machineId}": {
			"get": {
				"operationId": "GetAntivirusInformationByMachineId",
				"responses": {
					"200": {
						"description": "List",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IAntivirusInformation"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Antivirus Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateAntivirusInformation",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "string"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"tags": [
					"Antivirus Information"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_AntivirusInformation.machineId_"
							}
						}
					}
				}
			}
		},
...
		"/machines": {
			"get": {
				"operationId": "GetMachines",
				"responses": {
					"200": {
						"description": "List",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IMachine"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Get all registered machines (scoped to user's organization)",
				"tags": [
					"Machine"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateMachine",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "string"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Register a new machine",
				"tags": [
					"Machine"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Machine"
							}
						}
					}
				}
			}
		},
		"/machines/{id}": {
			"get": {
				"operationId": "GetMachine",
				"responses": {
					"200": {
						"description": "Machine",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IMachine"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Get a specific machine by database ID",
				"tags": [
					"Machine"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpsertMachine",
				"responses": {
					"200": {
						"description": "Updated"
					},
					"400": {
						"description": "Bad Request"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Update a machine",
				"tags": [
					"Machine"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Machine"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMachine",
				"responses": {
					"200": {
						"description": "Deleted"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Delete a machine",
				"tags": [
					"Machine"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/machines/by-machine-id/{machineId}": {
			"get": {
				"operationId": "GetMachineByMachineId",
				"responses": {
					"200": {
						"description": "Machine",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IMachine"
										}
									],
									"nullable": true
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"description": "Get a machine by machineId (client-generated ID)",
				"tags": [
					"Machine"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "machineId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}